ex 05

+++ src/main/java/sptech/com/br/exercicios/ex06/Questao02.java
@@ -42,7 +42,7 @@ public class Questao02 {
     }
 
     public static double avg(int[] integers) {
-        return sum(integers) / integers.length;
+        return sum(integers) / (double) integers.length;
     }
 
     public static int sum(int[] integers) {
